consolidate classes.
